Replace copy-files with public directories
Instead of listing files in a config block in package.json, we now pull them from the filesystem. The ember-electron directory essentially reflects the layout of the package we will assemble to run as the electron app, except it doesn't contain a package.json (we copy a modified version of the project root's package.json), it doesn't contain the built Ember app (obviously), and has several directories used to assemble platform-specific resources. The resources/ directory contains platform-independent files that are always included, and the resources-<platform> directories contain platform-specific files to augment or overwrite the platform-independent files.

So, with the exception of those missing files filled in my the build, and the special resources folders that are merged and pruned by the build, the ember-electron folder can be populated to look just like the build electron application package. Currently the main module is hard-coded to be at lib/index.js because electron-forge requires that the main module be in a subdirectory, and not at the root of the project.

This is all implemented as part of the broccoli pipeline now so we don't need special implementations for each different type of electron command.

When the ember electron command is run, the build output defaults to dist/ (mirroring ember serve). When the ember electron:package or ember electron:make commands are run, the packaged/make'd output is put in electron-out/.