Fix module load paths
This fixes #175 (which was actually an inability to require *any* dependencies). The problem was that we switched over to using a custom 'serve://' protocol to load index.html via electron-protocol-serve, but when loading index.html via a non-file protocol, electron's renderer/init.js doesn't know where it is in the filesystem, so it can't set up the node module search paths correctly.

So, this change adds some code to our shim-header.js to check to see if we were loaded by a non-file protocol, and if electron-protocol-serve set the ELECTRON_PROTOCOL_SERVE_INDEX environment variable giving us the path to index.html, and if so, runs some copypasta'ed code from electron's renderer/index.js to set up the module search paths correctly.