Revamp build pipeline
This is a significant re-write of the build pipeline, but it does not break compatibility. The blueprint is unaffected, and command functionality has been added, but not changed/removed. The only functional change (other than adding commands and command options) is that `ember electron` will build into `electron-out/project` and then run from there, rather than a temp directory, and leaves the output there when it exits. This matches `ember serve`'s behavior, and is nice for inspecting the files after running (especially if the app is crashing on startup!).

Code-architecturally, I moved us more in the direction of ember-cli by putting most of the work into tasks (and a couple of models), and making the commands mostly just a wrapper for the tasks (except I didn't really touch `ember electron` and `ember electron:test`). I also wrote a whole crap-ton of unit tests.

I added a build pipeline guide that describes how it works from a user/cli standpoint now, so I won't duplicate that here.