Fix & simplify package.json handling
Life is much much easier if our forge config is in the project's root package.json, so put it there via the blueprint, and stop putting it there dynamically in the build pipeline.

Also add a couple of checks to make sure things are in good shape during the build, addressing a few blueprint- and upgraded-related issues that have come up.

This fixes #232, fixes #236