Don't delete the app's node_modules
When running ember electron:test -s, anytime something was updated causing a rebuild, we'd delete a bunch of folders out of the application's node_modules folder...eep! The problem was that we weren't using a builder that was aware of the node_modules symlink, so when it synced the trees after a rebuild, it would think that most of the folders inside the symlinked node_modules shouldn't be there and would delete them.

So, refactor the builder/watcher code to make it usable by the test command, and also do a bit of cleanup to remove no-longer-needed code and make some stuff more readable.