Fix ember electron:test -s
I forgot to manually test this command since it would be pretty difficult to acceptance test, and it was broken :( The problem was that when running ember electron:test, the cleanup method of the builder is called before we run the tests, and it was deleting the node_modules symlink, and relying on the test-runner script to re-create it. But the whole point of this change was that when running ember electron:test -s, we leave the symlink in place while running the tests, so the test-runner was failing because the symlink already existed.

We don't need to delete the symlink in the builder's cleanup method -- it's not doing any harm, so we just leave it in the build output and don't mess with it in the test-runner script at all.