Fix packaging/making with npm/yarn
The code was running 'npm prune', but that doesn't do anything if no dependencies are already installed. What we really want is 'npm install --production' and I also added '--no-bin-links' to match the yarn behavior.

Additionally, we weren't using the same logic as ember to decide when to use yarn vs. npm. We would use yarn anytime it was present on the system, without checking to see if the project we are running in has a yarn.lock file. So, now we run the same logic to decide which to use, and set an environment variable to instruct electron-forge to match us.

Also, added end-to-end unit tests running with npm in addition to the ones running with yarn.

Fixes #308, #309, #312, #313