Do a minor overhaul of our electron:test setup
The goal here is to get rid of all the file hacking in test-runner.js, and have the brocolli pipeline manage all of our file operations. This involved moving the getTestemId() function that allow the testem client to communicate back to the server into a shim, and then doing some trickery in the main process to redirect the testem.js request to the testem server so we don't have to rewrite the script tag, which we can't do from broccoli because we don't know the testem URL yet.

In doing this, I realized that the tight coupling between test-runner.js and test-main.js is problematic because test-runner.js is used directly from the ember-electron package, while test-main.js is generated by the blueprint in the app project. So upgrades are hairy. So, this moves the test-main.js code out of the blueprint and into ember-electron so we can update it without requiring users to re-run the blueprint. If users want to customize test-main.js, they still can, but then they will be responsible for maintaining their project across updates to test-runner.js.

I also left the original test-runner.js in place (with a minor change to accomodate the assemble-tree.js change) so projects that use it, and therefore also have an old-style test-main.js, won't break when updating to this version of the addon.