Distinguish between "dev tools" and "inspector"
The term "inspector" is a little vague because of it could refer to either the browsers' "Inspect" menu command (i.e. open developer tools and show selected element) and the "Ember Inspector" (i.e. for examining details of the Ember application state). If I understand correctly, the current version of ember-electron strips devtron and the ember-inspector addons from the production build, but the dev tools are part of electron and never full removed (though can be `webPreferences: { devTools: false }`).