Set testem-electron.js in TestTask::run
This replaces the obsolete `testemOptions` method
while avoiding the need to use the newer
`defaultOptions` / `transformOptions` methods
(which won't work in ember-cli < 3.2).

Here we also drop testem from our dependencies
and instead let ember-cli use the version it wants.
(I'm not sure why we were bringing it in before.)
The breaking changes from 1.x to 2.0.0 were dropping
support for old things (Node 0.10 / 0.12 & IE8)
and the data sent with the `all-test-results` event.
Neither of these changes should affect this package.