fix: update semantic release commit message config
Rather than using eslint's config, this gets us on the default convention-changelog format. At least two maintainers of the project assumed this was the format (resulting in a release not happening because the format was wrong), plus it's aligned with the GitHub app we can install to validate the format for PRs.