chore: Misc updates (#402)
* fix: update semantic release commit message config

Rather than using eslint's config, this gets us on the default convention-changelog format. At least two maintainers of the project assumed this was the format (resulting in a release not happening because the format was wrong), plus it's aligned with the GitHub app we can install to validate the format for PRs.

* chore: remove package-lock.json

We don't really want two lock files to maintain, and @jacobq and I, the two most active developers on this project both use yarn. If this changes in the future, we can react accordingly.