fix: use a single version of electron-forge
We have two versions of electron-forge kicking around. ember-electron has it as a `dependency` so it can use it from the blueprint to create the Electron project using electron-forge's `init` command. When electron-forge creates a project, it installs itself in that project, so we end up with another version in `electron-app/package.json`.

When `ember-electron` was calling electron forge commands from `ember electron:*` commands, it was just `require()`ing electron-forge, which would get it the version in `/node_modules`. This could be a different version from the one in `electron-app/node_modules`, so the user could get different behavior running, e.g. `ember electron:make` vs. `cd electron-app && yarn electron-forge make`, breaking the simple mental model.

Also, recently electron-forge released a new beta version that changed the maker interface slighty, so the makers installed in `electron-app/` could be out of sync with the `electron-forge` version that the `ember electron` commands use. This is probably only a problem while electron-forge is in beta, but is also addressed by this change.

Now, other than the initial Electron project creation that happens from the blueprint, we only use the electron-forge in `electron-app`, which puts it right next to the makes in `package.json` allowing the user to easily see and manage the version without any hidden complexity.

Fixes #431