Change `vendor.js` to `app.js` in documentation
Because of how blunt `broccoli-asset-rev` is, the occurences of `vendor.js` in our documentation get rewritten with the fingerprinted filename `vendor-<hash>.js`, which is weird and confusing in the generated documentation site. So change to `app.js` which doesn't match any assets and won't get rewritten (and it just as good for the documentation/example purposes).