Un-pin electron forge
Probably as a leftover from when it was in beta, the `@electron-forge/core` dependency was pinned. Apparently `@electron-forge/template-base@6.0.5` doesn't work with `@electron-forge/core@6.0.4` (but core's dependency on template-base isn't pinned), so when installing `ember-electron` apps would end up with core v6.0.4 and template-base v6.0.5 and the blueprint wouldn't run. Un-pinning core should prevent this from happening again.

Fixes #1390